move label to separate component
