+++ airbyte-webapp/src/components/ServiceForm/components/Property/Label.tsx
@@ -1,70 +1,30 @@
-import React, { useMemo } from "react";
+import React from "react";
 import { useField } from "formik";
-import { FormattedMessage } from "react-intl";
 
-import TextWithHTML from "components/TextWithHTML";
 import { ControlLabels } from "components/LabeledControl";
 import { FormBaseItem } from "core/form/types";
 
-type IProps = {
+import { LabelMessage } from "./LabelMessage";
+
+type LabelMessageProps = {
   property: FormBaseItem;
 };
 
-const Label: React.FC<IProps> = ({ property, children }) => {
+const Label: React.FC<LabelMessageProps> = ({ property, children }) => {
   const [, meta] = useField(property.fieldName);
 
-  const label = `${property.title || property.fieldKey}${
-    property.isRequired ? " *" : ""
-  }`;
-
-  const constructExamples = () => {
-    if (!property.examples) {
-      return null;
-    }
-
-    const exampleText = Array.isArray(property.examples)
-      ? property.examples?.join(", ")
-      : property.examples;
-
-    return (
-      <FormattedMessage id="form.examples" values={{ examples: exampleText }} />
-    );
-  };
+  const labelText = property.title || property.fieldKey;
+  const labelRequiredAppendix = property.isRequired ? " *" : "";
+  const label = `${labelText}${labelRequiredAppendix}`;
 
   const displayError = !!meta.error && meta.touched;
-
-  const constructMessage = useMemo(() => {
-    const errorMessage =
-      displayError && meta.error === "form.pattern.error" ? (
-        <FormattedMessage
-          id={meta.error}
-          values={{ pattern: property.pattern }}
-        />
-      ) : null;
-
-    const message = property.description ? (
-      <TextWithHTML text={property.description} />
-    ) : null;
-
-    return (
-      <>
-        {message} {constructExamples()} {errorMessage}
-      </>
-    );
-  }, [
-    constructExamples,
-    displayError,
-    meta.error,
-    property.description,
-    property.pattern
-  ]);
 
   return (
     <ControlLabels
       labelAdditionLength={0}
       error={displayError}
       label={label}
-      message={constructMessage}
+      message={<LabelMessage property={property} />}
     >
       {children}
     </ControlLabels>
